+++ DZ/DZ/view_iPhone/source/A1_Newest_iphone.m
@@ -98,12 +98,12 @@ ON_SIGNAL3(newOrHotlistModel, RELOADED, signal)
     return  cell;
 }
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 { 
     return 90;
 }
 
--(int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+-(NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return self.newhotModel.shots.count;//forums.children.count;
 }
